requirements.txt
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
Brotli==1.0.9
certifi==2021.5.30
cffi==1.15.0
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distro==1.6.0
entrypoints==0.3
Flask==2.0.1
Flask-Compress==1.10.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.8.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
munch==2.5.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook~=6.4.8
numpy==1.21.1
packaging==21.0
pandas==1.3.1
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.21
pycparser==2.20
Pygments==2.10.0
pyparsing==3.0.1
pyproj==3.1.0
pyrsistent==0.18.0
PySimpleGUI==4.47.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.4
PyYAML==6.0
pyzmq==22.3.0
Send2Trash==1.8.0
Shapely==1.7.1
six==1.16.0
tabula-py==2.3.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1

streamlit~=1.5.0
matplotlib~=3.5.1
requests~=2.27.1
dotmap~=1.3.26
Pillow~=9.0.0
seaborn~=0.11.2
pyinstaller~=4.8
datapane~=0.13.2